backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
httpx>=0.24.0,<0.28
httpcore>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyjwt>=2.8.0

# Monetization
stripe>=7.0.0
apscheduler>=3.10.0
resend>=0.7.0

